Removed reliance on libm/math.h CRT functions.

+++ include/bx/inline/simd128_langext.inl
@@ -318,10 +318,10 @@ BX_SIMD128_IMPLEMENT_TEST(xyzw , 0xf);
 	BX_SIMD_FORCE_INLINE simd128_langext_t simd_sqrt(simd128_langext_t _a)
 	{
 		simd128_langext_t result;
-		result.vf[0] = sqrtf(_a.vf[0]);
-		result.vf[1] = sqrtf(_a.vf[1]);
-		result.vf[2] = sqrtf(_a.vf[2]);
-		result.vf[3] = sqrtf(_a.vf[3]);
+		result.vf[0] = sqrt(_a.vf[0]);
+		result.vf[1] = sqrt(_a.vf[1]);
+		result.vf[2] = sqrt(_a.vf[2]);
+		result.vf[3] = sqrt(_a.vf[3]);
 		return result;
 	}
 
@@ -329,10 +329,10 @@ BX_SIMD128_IMPLEMENT_TEST(xyzw , 0xf);
 	BX_SIMD_FORCE_INLINE simd128_langext_t simd_rsqrt_est(simd128_langext_t _a)
 	{
 		simd128_langext_t result;
-		result.vf[0] = 1.0f / sqrtf(_a.vf[0]);
-		result.vf[1] = 1.0f / sqrtf(_a.vf[1]);
-		result.vf[2] = 1.0f / sqrtf(_a.vf[2]);
-		result.vf[3] = 1.0f / sqrtf(_a.vf[3]);
+		result.vf[0] = 1.0f / sqrt(_a.vf[0]);
+		result.vf[1] = 1.0f / sqrt(_a.vf[1]);
+		result.vf[2] = 1.0f / sqrt(_a.vf[2]);
+		result.vf[3] = 1.0f / sqrt(_a.vf[3]);
 		return result;
 	}
